+++ src/main/java/com/space/server/auth/service/implementation/UserJoiner.java
@@ -4,8 +4,8 @@ import com.space.server.auth.presentation.dto.request.JoinUserRequest;
 import com.space.server.domin.inventory.domain.Inventory;
 import com.space.server.domin.inventory.service.implementation.InventoryCreator;
 import com.space.server.domin.inventory.service.implementation.InventoryUpdater;
-import com.space.server.core.item.domain.Item;
-import com.space.server.core.item.service.implementation.ItemReader;
+import com.space.server.domin.item.domain.Item;
+import com.space.server.domin.item.service.implementation.ItemReader;
 import com.space.server.user.domain.Users;
 import com.space.server.user.domain.repository.UserRepository;
 import com.space.server.user.domain.value.Role;
