+++ src/main/java/com/space/server/domin/inventory/domain/repository/InventoryRepository.java
@@ -1,8 +1,8 @@
 package com.space.server.domin.inventory.domain.repository;
 
 import com.space.server.domin.inventory.domain.Inventory;
-import com.space.server.core.item.domain.Item;
-import com.space.server.core.item.domain.value.Category;
+import com.space.server.domin.item.domain.Item;
+import com.space.server.domin.item.domain.value.Category;
 import com.space.server.user.domain.Users;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
