+++ src/main/java/com/space/server/domin/inventory/service/CommandInventoryService.java
@@ -1,9 +1,9 @@
 package com.space.server.domin.inventory.service;
 
 import com.space.server.domin.inventory.domain.Inventory;
-import com.space.server.core.item.domain.Item;
-import com.space.server.core.item.domain.value.Category;
-import com.space.server.core.item.service.implementation.ItemReader;
+import com.space.server.domin.item.domain.Item;
+import com.space.server.domin.item.domain.value.Category;
+import com.space.server.domin.item.service.implementation.ItemReader;
 import com.space.server.domin.inventory.service.implementation.*;
 import com.space.server.user.domain.Users;
 import com.space.server.user.domain.repository.UserRepository;
