+++ src/main/java/com/space/server/domin/inventory/service/implementation/InventoryValidator.java
@@ -2,7 +2,7 @@ package com.space.server.domin.inventory.service.implementation;
 
 import com.space.server.domin.inventory.domain.repository.InventoryRepository;
 import com.space.server.domin.inventory.exception.InventoryItemExistedException;
-import com.space.server.core.item.domain.Item;
+import com.space.server.domin.item.domain.Item;
 import com.space.server.user.domain.Users;
 import com.space.server.user.exception.InsufficientPointsException;
 import lombok.RequiredArgsConstructor;
