+++ src/main/java/com/space/server/domin/item/service/CommandItemService.java
@@ -1,10 +1,10 @@
-package com.space.server.core.item.service;
+package com.space.server.domin.item.service;
 
-import com.space.server.core.item.domain.Item;
-import com.space.server.core.item.service.implementation.ItemCreator;
-import com.space.server.core.item.service.implementation.ItemDeleter;
-import com.space.server.core.item.service.implementation.ItemReader;
-import com.space.server.core.item.service.implementation.ItemUpdater;
+import com.space.server.domin.item.domain.Item;
+import com.space.server.domin.item.service.implementation.ItemCreator;
+import com.space.server.domin.item.service.implementation.ItemDeleter;
+import com.space.server.domin.item.service.implementation.ItemReader;
+import com.space.server.domin.item.service.implementation.ItemUpdater;
 import org.springframework.transaction.annotation.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
