+++ src/main/java/com/space/server/domin/item/service/implementation/ItemReader.java
@@ -1,9 +1,9 @@
-package com.space.server.core.item.service.implementation;
+package com.space.server.domin.item.service.implementation;
 
-import com.space.server.core.item.domain.Item;
-import com.space.server.core.item.domain.value.Category;
-import com.space.server.core.item.domain.repository.ItemRepository;
-import com.space.server.core.item.exception.ItemNotFoundException;
+import com.space.server.domin.item.domain.Item;
+import com.space.server.domin.item.domain.value.Category;
+import com.space.server.domin.item.domain.repository.ItemRepository;
+import com.space.server.domin.item.exception.ItemNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
