+++ src/main/java/com/space/server/domin/item/service/implementation/ItemUpdater.java
@@ -1,6 +1,6 @@
-package com.space.server.core.item.service.implementation;
+package com.space.server.domin.item.service.implementation;
 
-import com.space.server.core.item.domain.Item;
+import com.space.server.domin.item.domain.Item;
 import org.springframework.stereotype.Service;
 
 @Service
